move testreportconversion to its own package and pull it's single use helpers out
