+++ main.go
@@ -20,7 +20,9 @@ import (
 	testgridv1 "github.com/openshift/sippy/pkg/apis/testgrid/v1"
 	"github.com/openshift/sippy/pkg/buganalysis"
 	"github.com/openshift/sippy/pkg/html"
+	"github.com/openshift/sippy/pkg/testgridanalysis"
 	"github.com/openshift/sippy/pkg/testgridanalysis/testgridanalysisapi"
+	"github.com/openshift/sippy/pkg/testgridanalysis/testreportconversion"
 	"github.com/openshift/sippy/pkg/util"
 	"github.com/openshift/sippy/pkg/util/sets"
 	"github.com/spf13/cobra"
@@ -294,7 +296,7 @@ func (a *Analyzer) processJobDetails(job testgridv1.JobDetails) {
 		test.Name = strings.TrimSpace(TagStripRegex.ReplaceAllString(test.Name, ""))
 		job.Tests[i] = test
 
-		a.processTest(job, platforms, test, util.FindSig(test.Name), startCol, endCol)
+		a.processTest(job, platforms, test, testgridanalysis.FindSig(test.Name), startCol, endCol)
 	}
 
 }
@@ -512,70 +514,22 @@ func downloadData(releases []string, filter string, storagePath string) {
 	}
 }
 
-// returns top ten failing tests w/o a bug and top ten with a bug(in that order)
-func getTopFailingTests(result map[string]sippyprocessingv1.SortedAggregateTestsResult, release string, bugCache buganalysis.BugCache) ([]*sippyprocessingv1.TestResult, []*sippyprocessingv1.TestResult) {
-	topTestsWithoutBug := []*sippyprocessingv1.TestResult{}
-	topTestsWithBug := []*sippyprocessingv1.TestResult{}
-	all := result["all"]
-	withoutbugcount := 0
-	withbugcount := 0
-	// look at the top 100 failing tests, try to create a list of the top 20 failures with bugs and without bugs.
-	// limit to 100 so we don't hammer search.ci too hard if we can't find 20 failures with bugs in the first 100.
-	for i := 0; (withbugcount < 20 || withoutbugcount < 10) && i < 100 && i < len(all.TestResults); i++ {
-		test := all.TestResults[i]
-		test.BugList = bugCache.ListBugs(release, "", test.Name)
-
-		// we want the top ten test failures that don't have bugs associated.
-		// top test failures w/ bugs will be listed, but don't count towards the top ten.
-		if len(test.BugList) == 0 && withoutbugcount < 10 {
-			topTestsWithoutBug = append(topTestsWithoutBug, &test)
-			withoutbugcount++
-		} else if len(test.BugList) > 0 && withbugcount < 20 {
-			topTestsWithBug = append(topTestsWithBug, &test)
-			withbugcount++
-		}
-	}
-	return topTestsWithoutBug, topTestsWithBug
-}
-
-func (a *Analyzer) prepareTestReport(prev bool) {
-	byAll := util.SummarizeTestResults(a.RawData.ByAll, a.BugCache, a.Release, a.Options.MinTestRuns, a.Options.TestSuccessThreshold)
-	byPlatform := util.SummarizeTestResults(a.RawData.ByPlatform, a.BugCache, a.Release, a.Options.MinTestRuns, a.Options.TestSuccessThreshold)
-	byJob := util.SummarizeTestResults(a.RawData.ByJob, a.BugCache, a.Release, a.Options.MinTestRuns, a.Options.TestSuccessThreshold)
-	bySig := util.SummarizeTestResults(a.RawData.BySig, a.BugCache, a.Release, a.Options.MinTestRuns, a.Options.TestSuccessThreshold)
-
-	filteredFailureGroups := util.FilterFailureGroups(a.RawData.JobResults, a.BugCache, a.Release, a.Options.FailureClusterThreshold)
-	jobResults, infrequentJobResults := util.SummarizeJobRunResults(a.RawData.JobResults, byJob, a.BugCache, a.Release, a.Options.EndDay)
-
-	bugFailureCounts := util.GenerateSortedBugFailureCounts(a.RawData.JobResults, byAll, a.BugCache, a.Release)
-	bugzillaComponentResults := util.GenerateJobFailuresByBugzillaComponent(a.RawData.JobResults, byJob)
-
-	a.Report = sippyprocessingv1.TestReport{
-		Release:                        a.Release,
-		All:                            byAll,
-		ByPlatform:                     byPlatform,
-		ByJob:                          byJob,
-		BySig:                          bySig,
-		FailureGroups:                  filteredFailureGroups,
-		JobResults:                     jobResults,
-		InfrequentJobResults:           infrequentJobResults,
-		Timestamp:                      a.LastUpdateTime,
-		BugsByFailureCount:             bugFailureCounts,
-		JobFailuresByBugzillaComponent: bugzillaComponentResults,
-
-		AnalysisWarnings: a.analysisWarnings,
-	}
-
-	if !prev {
-		topFailingTestsWithoutBug, topFailingTestsWithBug := getTopFailingTests(byAll, a.Release, a.BugCache)
-		a.Report.TopFailingTestsWithBug = topFailingTestsWithBug
-		a.Report.TopFailingTestsWithoutBug = topFailingTestsWithoutBug
-	}
-
+func (a *Analyzer) prepareTestReport() {
+	a.Report = testreportconversion.PrepareTestReport(
+		a.RawData,
+		a.BugCache,
+		a.Release,
+		a.Options.MinTestRuns,
+		a.Options.TestSuccessThreshold,
+		a.Options.EndDay,
+		a.analysisWarnings,
+		a.LastUpdateTime,
+		a.Options.FailureClusterThreshold,
+	)
 }
 
 func (a *Analyzer) printReport() {
-	a.prepareTestReport(false)
+	a.prepareTestReport()
 	switch a.Options.Output {
 	case "json":
 		a.printJsonReport()
@@ -781,7 +735,7 @@ func (s *Server) refresh(w http.ResponseWriter, req *http.Request) {
 
 		analyzer.loadData([]string{analyzer.Release}, analyzer.Options.LocalData)
 		analyzer.analyze()
-		analyzer.prepareTestReport(strings.Contains(k, "-prev"))
+		analyzer.prepareTestReport()
 		s.analyzers[k] = analyzer
 	}
 
@@ -904,7 +858,7 @@ func (s *Server) detailed(w http.ResponseWriter, req *http.Request) {
 	}
 	analyzer.loadData([]string{release}, s.options.LocalData)
 	analyzer.analyze()
-	analyzer.prepareTestReport(false)
+	analyzer.prepareTestReport()
 
 	// prior 7 day period
 	optCopy := *opt
@@ -924,7 +878,7 @@ func (s *Server) detailed(w http.ResponseWriter, req *http.Request) {
 	}
 	prevAnalyzer.loadData([]string{release}, s.options.LocalData)
 	prevAnalyzer.analyze()
-	prevAnalyzer.prepareTestReport(true)
+	prevAnalyzer.prepareTestReport()
 
 	html.PrintHtmlReport(w, req, analyzer.Report, prevAnalyzer.Report, opt.EndDay, jobTestCount)
 
@@ -1053,7 +1007,7 @@ func (o *Options) Run() error {
 			}
 			analyzer.loadData([]string{release}, o.LocalData)
 			analyzer.analyze()
-			analyzer.prepareTestReport(false)
+			analyzer.prepareTestReport()
 			server.analyzers[release] = analyzer
 
 			// prior 7 day period (days 7-14)
@@ -1074,7 +1028,7 @@ func (o *Options) Run() error {
 			}
 			analyzer.loadData([]string{release}, o.LocalData)
 			analyzer.analyze()
-			analyzer.prepareTestReport(true)
+			analyzer.prepareTestReport()
 			server.analyzers[release+"-prev"] = analyzer
 		}
 		server.serve(o)
