+++ pkg/util/utils.go
@@ -1,25 +1,18 @@
 package util
 
 import (
-	"fmt"
-	gohtml "html"
 	"math"
 	"regexp"
 	"sort"
 	"strings"
 	"time"
 
-	bugsv1 "github.com/openshift/sippy/pkg/apis/bugs/v1"
 	sippyprocessingv1 "github.com/openshift/sippy/pkg/apis/sippyprocessing/v1"
-	"github.com/openshift/sippy/pkg/buganalysis"
-	"github.com/openshift/sippy/pkg/testgridanalysis"
 	"github.com/openshift/sippy/pkg/testgridanalysis/testgridanalysisapi"
-	"github.com/openshift/sippy/pkg/util/sets"
 	"k8s.io/klog"
 )
 
 var (
-	sigRegex      *regexp.Regexp = regexp.MustCompile(`\[(sig-.*?)\]`)
 	bugzillaRegex *regexp.Regexp = regexp.MustCompile(`(https://bugzilla.redhat.com/show_bug.cgi\?id=\d+)`)
 
 	// platform regexes
@@ -102,221 +95,6 @@ func Percent(success, failure int) float64 {
 	return float64(success) / float64(success+failure) * 100.0
 }
 
-func SummarizeTestResults(
-	aggregateTestResult map[string]testgridanalysisapi.AggregateTestsResult,
-	bugCache buganalysis.BugCache, // required to associate tests with bug
-	release string, // required to limit bugs to those that apply to the release in question
-	minRuns int, // indicates how many runs are required for a test is included in overall percentages
-	// TODO deads2k wants to eliminate the successThreshold
-	successThreshold float64, // indicates an upper bound on how successful a test can be before it is excluded
-) map[string]sippyprocessingv1.SortedAggregateTestsResult {
-	sorted := make(map[string]sippyprocessingv1.SortedAggregateTestsResult)
-
-	for k, v := range aggregateTestResult {
-		sorted[k] = sippyprocessingv1.SortedAggregateTestsResult{}
-
-		passedCount := 0
-		failedCount := 0
-		for _, rawTestResult := range v.RawTestResults {
-			passPercentage := Percent(rawTestResult.Successes, rawTestResult.Failures)
-
-			// strip out tests are more than N% successful
-			if passPercentage > successThreshold {
-				continue
-			}
-			// strip out tests that have less than N total runs
-			if rawTestResult.Successes+rawTestResult.Failures < minRuns {
-				continue
-			}
-
-			passedCount += rawTestResult.Successes
-			failedCount += rawTestResult.Failures
-
-			testSearchUrl := gohtml.EscapeString(regexp.QuoteMeta(rawTestResult.Name))
-			testSearchLink := fmt.Sprintf("<a target=\"_blank\" href=\"https://search.ci.openshift.org/?maxAge=48h&context=1&type=bug%%2Bjunit&name=&maxMatches=5&maxBytes=20971520&groupBy=job&search=%s\">%s</a>", testSearchUrl, rawTestResult.Name)
-
-			s := sorted[k]
-			s.TestResults = append(s.TestResults, sippyprocessingv1.TestResult{
-				Name:           rawTestResult.Name,
-				Successes:      rawTestResult.Successes,
-				Failures:       rawTestResult.Failures,
-				Flakes:         rawTestResult.Flakes,
-				PassPercentage: passPercentage,
-				BugList:        bugCache.ListBugs(release, "", rawTestResult.Name),
-				SearchLink:     testSearchLink,
-			})
-			sorted[k] = s
-		}
-
-		s := sorted[k]
-		s.Successes = passedCount
-		s.Failures = failedCount
-		s.TestPassPercentage = Percent(passedCount, failedCount)
-		sorted[k] = s
-
-		// sort from lowest to highest
-		sort.SliceStable(sorted[k].TestResults, func(i, j int) bool {
-			return sorted[k].TestResults[i].PassPercentage < sorted[k].TestResults[j].PassPercentage
-		})
-	}
-	return sorted
-}
-
-func GenerateSortedBugFailureCounts(
-	allJobResults map[string]testgridanalysisapi.RawJobResult,
-	byAll map[string]sippyprocessingv1.SortedAggregateTestsResult,
-	bugCache buganalysis.BugCache, // required to associate tests with bug
-	release string, // required to limit bugs to those that apply to the release in question
-) []bugsv1.Bug {
-	bugs := map[string]bugsv1.Bug{}
-
-	failedTestNamesAcrossAllJobRuns := sets.NewString()
-	for _, jobResult := range allJobResults {
-		for _, jobrun := range jobResult.JobRunResults {
-			failedTestNamesAcrossAllJobRuns.Insert(jobrun.FailedTestNames...)
-		}
-	}
-
-	// for every test that failed in some job run, look up the bug(s) associated w/ the test
-	// and attribute the number of times the test failed+flaked to that bug(s)
-	for _, testResult := range byAll["all"].TestResults {
-		testName := testResult.Name
-		bugList := bugCache.ListBugs(release, "", testName)
-		for _, bug := range bugList {
-			if b, found := bugs[bug.Url]; found {
-				b.FailureCount += testResult.Failures
-				b.FlakeCount += testResult.Flakes
-				bugs[bug.Url] = b
-			} else {
-				bug.FailureCount = testResult.Failures
-				bug.FlakeCount = testResult.Flakes
-				bugs[bug.Url] = bug
-			}
-		}
-	}
-
-	sortedBugs := []bugsv1.Bug{}
-	for _, bug := range bugs {
-		sortedBugs = append(sortedBugs, bug)
-	}
-	// sort from highest to lowest
-	sort.SliceStable(sortedBugs, func(i, j int) bool {
-		return sortedBugs[i].FailureCount > sortedBugs[j].FailureCount
-	})
-	return sortedBugs
-}
-
-func FilterFailureGroups(
-	rawJobResults map[string]testgridanalysisapi.RawJobResult,
-	bugCache buganalysis.BugCache, // required to associate tests with bug
-	release string, // required to limit bugs to those that apply to the release in question
-	failureClusterThreshold int,
-) []sippyprocessingv1.JobRunResult {
-	filteredJrr := []sippyprocessingv1.JobRunResult{}
-	// -1 means don't do this reporting.
-	if failureClusterThreshold < 0 {
-		return filteredJrr
-	}
-	for _, jobResult := range rawJobResults {
-		for _, rawJRR := range jobResult.JobRunResults {
-			if rawJRR.TestFailures < failureClusterThreshold {
-				continue
-			}
-
-			allFailuresKnown := areAllFailuresKnown(rawJRR, bugCache, release)
-			hasUnknownFailure := rawJRR.Failed && !allFailuresKnown
-
-			filteredJrr = append(filteredJrr, sippyprocessingv1.JobRunResult{
-				Job:                jobResult.JobName,
-				Url:                rawJRR.JobRunURL,
-				TestGridJobUrl:     jobResult.TestGridJobUrl,
-				TestFailures:       rawJRR.TestFailures,
-				FailedTestNames:    rawJRR.FailedTestNames,
-				Failed:             rawJRR.Failed,
-				HasUnknownFailures: hasUnknownFailure,
-				Succeeded:          rawJRR.Succeeded,
-			})
-		}
-	}
-
-	// sort from highest to lowest
-	sort.SliceStable(filteredJrr, func(i, j int) bool {
-		return filteredJrr[i].TestFailures > filteredJrr[j].TestFailures
-	})
-
-	return filteredJrr
-}
-
-func areAllFailuresKnown(
-	rawJRR testgridanalysisapi.RawJobRunResult,
-	bugCache buganalysis.BugCache, // required to associate tests with bug
-	release string, // required to limit bugs to those that apply to the release in question,
-) bool {
-	// check if all the test failures in the run can be attributed to
-	// known bugs.  If not, the job run was an "unknown failure" that we cannot pretend
-	// would have passed if all our bugs were fixed.
-	allFailuresKnown := true
-	for _, testName := range rawJRR.FailedTestNames {
-		bugs := bugCache.ListBugs(release, "", testName)
-		isKnownFailure := len(bugs) > 0
-		if !isKnownFailure {
-			allFailuresKnown = false
-			break
-		}
-	}
-	return allFailuresKnown
-}
-
-func SummarizeJobRunResults(
-	rawJobResults map[string]testgridanalysisapi.RawJobResult,
-	byJob map[string]sippyprocessingv1.SortedAggregateTestsResult,
-	bugCache buganalysis.BugCache, // required to associate tests with bug
-	release string, // required to limit bugs to those that apply to the release in question,
-	numberOfDaysOfData int, // number of days included in report.
-) (jobs []sippyprocessingv1.JobResult, infrequentJobs []sippyprocessingv1.JobResult) {
-
-	for jobName, rawJobResult := range rawJobResults {
-		job := sippyprocessingv1.JobResult{
-			Name:        jobName,
-			TestResults: byJob[jobName].TestResults,
-		}
-
-		for _, rawJRR := range rawJobResult.JobRunResults {
-			// TODO move into RawJobResult
-			job.TestGridUrl = rawJobResult.TestGridJobUrl
-
-			if rawJRR.Failed {
-				job.Failures++
-			} else if rawJRR.Succeeded {
-				job.Successes++
-			}
-			if rawJRR.Failed && areAllFailuresKnown(rawJRR, bugCache, release) {
-				job.KnownFailures++
-			}
-		}
-
-		job.PassPercentage = Percent(job.Successes, job.Failures)
-		job.PassPercentageWithKnownFailures = Percent(job.Successes+job.KnownFailures, job.Failures-job.KnownFailures)
-
-		if job.Successes+job.Failures > numberOfDaysOfData*3/2 /*time 1.5*/ {
-			jobs = append(jobs, job)
-		} else {
-			infrequentJobs = append(infrequentJobs, job)
-		}
-	}
-
-	// sort from lowest to highest
-	sort.SliceStable(jobs, func(i, j int) bool {
-		return jobs[i].PassPercentage < jobs[j].PassPercentage
-	})
-	// sort from lowest to highest
-	sort.SliceStable(infrequentJobs, func(i, j int) bool {
-		return infrequentJobs[i].PassPercentage < infrequentJobs[j].PassPercentage
-	})
-
-	return jobs, infrequentJobs
-}
-
 func RelevantJob(jobName, status string, filter *regexp.Regexp) bool {
 	if filter != nil && !filter.MatchString(jobName) {
 		return false
@@ -346,15 +124,6 @@ func ComputeLookback(startday, lookback int, timestamps []int) (int, int) {
 		}
 	}
 	return start, len(timestamps)
-}
-
-// find associated sig from test name
-func FindSig(name string) string {
-	match := sigRegex.FindStringSubmatch(name)
-	if len(match) > 1 {
-		return match[1]
-	}
-	return "sig-unknown"
 }
 
 func FindPlatform(name string) []string {
@@ -510,156 +279,4 @@ func GetPrevBugzillaJobFailures(bzComponent string, bugzillaJobFailures []sippyp
 		}
 	}
 	return nil
-}
-
-func GenerateJobFailuresByBugzillaComponent(
-	allJobResults map[string]testgridanalysisapi.RawJobResult,
-	jobToTestResults map[string]sippyprocessingv1.SortedAggregateTestsResult,
-) map[string]sippyprocessingv1.SortedBugzillaComponentResult {
-
-	bzComponentToBZJobResults := map[string][]sippyprocessingv1.BugzillaJobResult{}
-	for job, jobResult := range allJobResults {
-		curr := generateJobFailuresByBugzillaComponent(job, jobResult.JobRunResults, jobToTestResults[job].TestResults)
-		// each job will be distinct, so we merely need to append
-		for bzComponent, bzJobResult := range curr {
-			bzComponentToBZJobResults[bzComponent] = append(bzComponentToBZJobResults[bzComponent], bzJobResult)
-		}
-	}
-
-	sortedResults := map[string]sippyprocessingv1.SortedBugzillaComponentResult{}
-	for bzComponent, jobResults := range bzComponentToBZJobResults {
-		// sort from least passing to most passing
-		// we expect these lists to be small, so the sort isn't awful
-		sort.SliceStable(jobResults, func(i, j int) bool {
-			return jobResults[i].FailPercentage > jobResults[j].FailPercentage
-		})
-		sortedResults[bzComponent] = sippyprocessingv1.SortedBugzillaComponentResult{
-			Name:       bzComponent,
-			JobsFailed: jobResults,
-		}
-	}
-
-	return sortedResults
-}
-
-// returns bz component to bzJob
-func generateJobFailuresByBugzillaComponent(
-	jobName string,
-	jobRuns map[string]testgridanalysisapi.RawJobRunResult,
-	jobTestResults []sippyprocessingv1.TestResult,
-) map[string]sippyprocessingv1.BugzillaJobResult {
-
-	bzComponentToFailedJobRuns := map[string]sets.String{}
-	bzToTestNameToTestResult := map[string]map[string]sippyprocessingv1.TestResult{}
-	failedTestCount := 0
-	foundTestCount := 0
-	for _, rawJRR := range jobRuns {
-		failedTestCount += len(rawJRR.FailedTestNames)
-		for _, testName := range rawJRR.FailedTestNames {
-			testResult, foundTest := getTestResultForJob(jobTestResults, testName)
-			if !foundTest {
-				continue
-			}
-			foundTestCount++
-
-			bzComponents := getBugzillaComponentsFromTestResult(testResult)
-			for _, bzComponent := range bzComponents {
-				// set the failed runs so we know which jobs failed
-				failedJobRuns, ok := bzComponentToFailedJobRuns[bzComponent]
-				if !ok {
-					failedJobRuns = sets.String{}
-				}
-				failedJobRuns.Insert(rawJRR.JobRunURL)
-				bzComponentToFailedJobRuns[bzComponent] = failedJobRuns
-				////////////////////////////////
-
-				// set the filtered test result
-				testNameToTestResult, ok := bzToTestNameToTestResult[bzComponent]
-				if !ok {
-					testNameToTestResult = map[string]sippyprocessingv1.TestResult{}
-				}
-				testNameToTestResult[testName] = getTestResultFilteredByComponent(testResult, bzComponent)
-				bzToTestNameToTestResult[bzComponent] = testNameToTestResult
-				////////////////////////////////
-			}
-		}
-	}
-
-	bzComponentToBZJobResult := map[string]sippyprocessingv1.BugzillaJobResult{}
-	for bzComponent, failedJobRuns := range bzComponentToFailedJobRuns {
-		totalRuns := len(jobRuns)
-		numFailedJobRuns := len(failedJobRuns)
-		failPercentage := float64(numFailedJobRuns*100) / float64(totalRuns)
-
-		bzJobTestResult := []sippyprocessingv1.TestResult{}
-		for _, testResult := range bzToTestNameToTestResult[bzComponent] {
-			bzJobTestResult = append(bzJobTestResult, testResult)
-		}
-		// sort from least passing to most passing
-		sort.SliceStable(bzJobTestResult, func(i, j int) bool {
-			return bzJobTestResult[i].PassPercentage < bzJobTestResult[j].PassPercentage
-		})
-
-		bzComponentToBZJobResult[bzComponent] = sippyprocessingv1.BugzillaJobResult{
-			JobName:               jobName,
-			BugzillaComponent:     bzComponent,
-			NumberOfJobRunsFailed: numFailedJobRuns,
-			FailPercentage:        failPercentage,
-			TotalRuns:             totalRuns,
-			Failures:              bzJobTestResult,
-		}
-	}
-
-	return bzComponentToBZJobResult
-}
-
-func getBugzillaComponentsFromTestResult(testResult sippyprocessingv1.TestResult) []string {
-	bzComponents := sets.String{}
-	bugList := testResult.BugList
-	for _, bug := range bugList {
-		bzComponents.Insert(bug.Component[0])
-	}
-	if len(bzComponents) > 0 {
-		return bzComponents.List()
-	}
-
-	// If we didn't have a bug, use the test name itself to identify a likely victim/blame
-	switch {
-	case strings.HasPrefix(testResult.Name, testgridanalysisapi.OperatorInstallPrefix):
-		operatorName := testResult.Name[len(testgridanalysisapi.OperatorInstallPrefix):]
-		return []string{testgridanalysis.GetBugzillaComponentForOperator(operatorName)}
-
-	case strings.HasPrefix(testResult.Name, testgridanalysisapi.OperatorUpgradePrefix):
-		operatorName := testResult.Name[len(testgridanalysisapi.OperatorUpgradePrefix):]
-		return []string{testgridanalysis.GetBugzillaComponentForOperator(operatorName)}
-
-	default:
-		return []string{testgridanalysis.GetBugzillaComponentForSig(FindSig(testResult.Name))}
-	}
-
-}
-
-func getTestResultForJob(jobTestResults []sippyprocessingv1.TestResult, testName string) (sippyprocessingv1.TestResult, bool) {
-	for _, testResult := range jobTestResults {
-		if testResult.Name == testName {
-			return testResult, true
-		}
-	}
-	return sippyprocessingv1.TestResult{
-		Name:           "if-seen-report-bug---" + testName,
-		PassPercentage: 200.0,
-	}, false
-}
-
-func getTestResultFilteredByComponent(testResult sippyprocessingv1.TestResult, bzComponent string) sippyprocessingv1.TestResult {
-	ret := testResult
-	ret.BugList = []bugsv1.Bug{}
-	for i := range testResult.BugList {
-		bug := testResult.BugList[i]
-		if bug.Component[0] == bzComponent {
-			ret.BugList = append(ret.BugList, bug)
-		}
-	}
-
-	return ret
 }
